protect et overflow

+++ philo_bonus/srcs/ft_finish_bonus.c
@@ -6,7 +6,7 @@
 /*   By: nflan <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/20 11:31:57 by nflan             #+#    #+#             */
-/*   Updated: 2022/05/10 15:21:50 by nflan            ###   ########.fr       */
+/*   Updated: 2022/05/25 15:23:59 by nflan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	*ft_death_checker(void *arg)
 	t_phil	*philo;
 
 	philo = (t_phil *)arg;
-	usleep(philo->g->tdie * 900);
+	ft_usleep(philo->g->tdie * 900);
 	while (1)
 	{
 		sem_wait(philo->g->die);
@@ -45,13 +45,21 @@ void	ft_end_philo(t_all *g, t_phil *philo)
 	int	i;
 
 	i = -1;
-	while (++i < g->nbphilo)
-		if (philo[i].pid != -1)
-			kill(philo[i].pid, SIGKILL);
-	sem_close(g->forks);
-	sem_close(g->death);
-	sem_close(g->print);
-	sem_close(g->take);
-	sem_close(g->die);
-	free(philo);
+	if (philo)
+	{
+		while (++i < g->nbphilo)
+			if (philo[i].pid != -1)
+				kill(philo[i].pid, SIGKILL);
+		free(philo);
+	}
+	if (g->forks)
+		sem_close(g->forks);
+	if (g->death)
+		sem_close(g->death);
+	if (g->print)
+		sem_close(g->print);
+	if (g->take)
+		sem_close(g->take);
+	if (g->die)
+		sem_close(g->die);
 }
